support upgrade target versions
