+++ config.js
@@ -11,6 +11,7 @@ const MANIFEST_OUT_DIR = path.join("manifest");
 
 export default {
   constants: {
+    CROWDBOTICS_FILE: ".crowdbotics.json",
     COOKIECUTTER_PACKAGE: "cookiecutter==1.7.3",
     REACT_NATIVE_SCAFFOLD_REPO_ORIGIN:
       "https://github.com/crowdbotics/react-native-scaffold"
